Update pyproject.toml to fix issue #2286
Fixed a bug in oscrypto where the regex would fail to recognize the right version of SSL due to insufficient check on multiple digits. 
However the owner of the repo does not plan to make a new release soon due to time constraints. 
See -> https://github.com/wbond/oscrypto/issues/78 

This can be fixed by using a specific version from git pointing to the right commit. 
Hence this change in the pyproject.toml file.